[#247] feat: integrate new container system - Part III - add KActivator2 to engine context
