+++ implementations/implementations-std-core/src/test/java/io/github/darthakiranihil/konna/core/message/KStandardMessengerPositiveTests.java
@@ -122,8 +122,8 @@ public class KStandardMessengerPositiveTests extends KStandardTestClass {
 
     @Test
     public void testSendRegularMessageSuccess() {
-        KMessenger messenger1 = new KStandardMessenger(this.messageSystem, "TestComponent");
-        KMessenger messenger2 = new KStandardMessenger(this.messageSystem, "TestComponentAgain");
+        KMessenger messenger1 = new KStandardMessenger(this.messageSystem);
+        KMessenger messenger2 = new KStandardMessenger(this.messageSystem);
 
         this.sendTest(
             messenger1::sendRegularSync,
@@ -134,8 +134,8 @@ public class KStandardMessengerPositiveTests extends KStandardTestClass {
 
     @Test
     public void testSendDebugMessageSuccess() {
-        KMessenger messenger1 = new KStandardMessenger(this.messageSystem, "TestComponent");
-        KMessenger messenger2 = new KStandardMessenger(this.messageSystem, "TestComponentAgain");
+        KMessenger messenger1 = new KStandardMessenger(this.messageSystem);
+        KMessenger messenger2 = new KStandardMessenger(this.messageSystem);
 
         this.sendTest(
             messenger1::sendDebugSync,
@@ -146,8 +146,8 @@ public class KStandardMessengerPositiveTests extends KStandardTestClass {
 
     @Test
     public void testSendMetricsMessageSuccess() {
-        KMessenger messenger1 = new KStandardMessenger(this.messageSystem, "TestComponent");
-        KMessenger messenger2 = new KStandardMessenger(this.messageSystem, "TestComponentAgain");
+        KMessenger messenger1 = new KStandardMessenger(this.messageSystem);
+        KMessenger messenger2 = new KStandardMessenger(this.messageSystem);
 
         this.sendTest(
             messenger1::sendMetricsSync,
@@ -158,8 +158,8 @@ public class KStandardMessengerPositiveTests extends KStandardTestClass {
 
     @Test
     public void testSendSystemMessageSuccess() {
-        KMessenger messenger1 = new KStandardMessenger(this.messageSystem, "TestComponent");
-        KMessenger messenger2 = new KStandardMessenger(this.messageSystem, "TestComponentAgain");
+        KMessenger messenger1 = new KStandardMessenger(this.messageSystem);
+        KMessenger messenger2 = new KStandardMessenger(this.messageSystem);
 
         this.sendTest(
             messenger1::sendSystemSync,
@@ -170,8 +170,8 @@ public class KStandardMessengerPositiveTests extends KStandardTestClass {
 
     @Test
     public void testSendRegularMessageAsyncSuccess() {
-        KMessenger messenger1 = new KStandardMessenger(this.messageSystem, "TestComponent");
-        KMessenger messenger2 = new KStandardMessenger(this.messageSystem, "TestComponentAgain");
+        KMessenger messenger1 = new KStandardMessenger(this.messageSystem);
+        KMessenger messenger2 = new KStandardMessenger(this.messageSystem);
 
         this.sendTest(
             messenger1::sendRegular,
@@ -182,8 +182,8 @@ public class KStandardMessengerPositiveTests extends KStandardTestClass {
 
     @Test
     public void testSendDebugMessageAsyncSuccess() {
-        KMessenger messenger1 = new KStandardMessenger(this.messageSystem, "TestComponent");
-        KMessenger messenger2 = new KStandardMessenger(this.messageSystem, "TestComponentAgain");
+        KMessenger messenger1 = new KStandardMessenger(this.messageSystem);
+        KMessenger messenger2 = new KStandardMessenger(this.messageSystem);
 
         this.sendTest(
             messenger1::sendDebug,
@@ -194,8 +194,8 @@ public class KStandardMessengerPositiveTests extends KStandardTestClass {
 
     @Test
     public void testSendMetricsMessageAsyncSuccess() {
-        KMessenger messenger1 = new KStandardMessenger(this.messageSystem, "TestComponent");
-        KMessenger messenger2 = new KStandardMessenger(this.messageSystem, "TestComponentAgain");
+        KMessenger messenger1 = new KStandardMessenger(this.messageSystem);
+        KMessenger messenger2 = new KStandardMessenger(this.messageSystem);
 
         this.sendTest(
             messenger1::sendMetrics,
@@ -206,8 +206,8 @@ public class KStandardMessengerPositiveTests extends KStandardTestClass {
 
     @Test
     public void testSendSystemMessageAsyncSuccess() {
-        KMessenger messenger1 = new KStandardMessenger(this.messageSystem, "TestComponent");
-        KMessenger messenger2 = new KStandardMessenger(this.messageSystem, "TestComponentAgain");
+        KMessenger messenger1 = new KStandardMessenger(this.messageSystem);
+        KMessenger messenger2 = new KStandardMessenger(this.messageSystem);
 
         this.sendTest(
             messenger1::sendSystem,
