+++ testing/src/main/java/io/github/darthakiranihil/konna/test/KStandardTestClass.java
@@ -16,32 +16,15 @@
 
 package io.github.darthakiranihil.konna.test;
 
-import io.github.darthakiranihil.konna.core.app.*;
 import io.github.darthakiranihil.konna.core.data.json.*;
 import io.github.darthakiranihil.konna.core.di.KContainerModifier;
-import io.github.darthakiranihil.konna.core.di.KStandardContainerAccessor;
 import io.github.darthakiranihil.konna.core.engine.KEngineContext;
-import io.github.darthakiranihil.konna.core.engine.KProxiedEngineContext;
-import io.github.darthakiranihil.konna.core.io.KAssetLoader;
-import io.github.darthakiranihil.konna.core.io.KJsonSubtypeBasedAssetLoader;
-import io.github.darthakiranihil.konna.core.io.KStandardResourceLoader;
-import io.github.darthakiranihil.konna.core.io.protocol.KClasspathProtocol;
-import io.github.darthakiranihil.konna.core.log.KLogLevel;
-import io.github.darthakiranihil.konna.core.log.KLogger;
-import io.github.darthakiranihil.konna.core.log.KSimpleLogFormatter;
-import io.github.darthakiranihil.konna.core.log.system.*;
-import io.github.darthakiranihil.konna.core.message.*;
-import io.github.darthakiranihil.konna.core.object.*;
+import io.github.darthakiranihil.konna.core.message.KQueueBasedMessageSystem;
+import io.github.darthakiranihil.konna.core.object.KObject;
+import io.github.darthakiranihil.konna.core.object.KTag;
 import io.github.darthakiranihil.konna.core.struct.KStructUtils;
-import io.github.darthakiranihil.konna.core.util.KCache;
-import io.github.darthakiranihil.konna.core.util.KClassGraphClasspathSearchEngine;
-import io.github.darthakiranihil.konna.core.util.KClasspathSearchEngine;
-import io.github.darthakiranihil.konna.core.util.KHashMapBasedCache;
 import org.jetbrains.annotations.TestOnly;
 import org.jspecify.annotations.Nullable;
-
-import java.util.List;
-import java.util.Map;
 
 /**
  * Standard test class, containing implementations of most common Konna classes.
@@ -91,69 +74,6 @@ public class KStandardTestClass extends KObject {
      * Convenience reference to message system.
      */
     protected static @Nullable KQueueBasedMessageSystem msgSystem;
-
-    static {
-        var classpath = new KClassGraphClasspathSearchEngine();
-        var containerResolver = new KStandardContainerAccessor();
-        containerResolver
-            .getContainer()
-            .add(KJsonParser.class, KStandardJsonParser.class)
-            .add(KJsonTokenizer.class, KStandardJsonTokenizer.class)
-            .add(KJsonDeserializer.class, KStandardJsonDeserializer.class)
-            .add(KJsonSerializer.class, KStandardJsonSerializer.class)
-            .add(KActivator.class, KProxiedEngineContext.class)
-            .add(KMessageSystem.class, KProxiedEngineContext.class)
-            .add(KAssetLoader.class, KProxiedEngineContext.class)
-            .add(KEventSystem.class, KProxiedEngineContext.class)
-            .add(KMessenger.class, KStandardMessenger.class)
-            .add(KLogger.class, KStandardLogger.class)
-            .add(KClasspathSearchEngine.class, KClassGraphClasspathSearchEngine.class)
-            .add(KFrameTaskScheduler.class, KProxiedEngineContext.class)
-            .add(KFrameTaskExecutor.class, KStandardFrameTaskSystem.class)
-            .add(KFrameTaskSystem.class, KStandardFrameTaskSystem.class)
-            .add(KFrameTaskPrioritizer.class, KFrameTaskPrioritizer.LeaveAsIs.class)
-            .add(KCache.class, KHashMapBasedCache.class);
-
-        var objectRegistry = new KStandardObjectRegistry();
-        var activator = new KStandardActivator(containerResolver, objectRegistry, classpath);
-        var messageSystem = new KStandardMessageSystem(activator);
-        var eventSystem = new KStandardEventSystem();
-        var resourceLoader = new KStandardResourceLoader(
-            List.of(
-                new KClasspathProtocol(
-                    ClassLoader.getSystemClassLoader()
-                )
-            )
-        );
-        var assetLoader = new KJsonSubtypeBasedAssetLoader(
-            resourceLoader,
-            Map.of(),
-            new KStandardJsonParser(new KStandardJsonTokenizer())
-        );
-        var frameTaskSystem = activator.createObject(KFrameTaskSystem.class);
-        KSystemLogger.addLogHandler(new KFileLogHandler("_log.log", new KTimestampLogFormatter()));
-        KSystemLogger.addLogHandler(new KTerminalLogHandler(new KColorfulTerminalLogFormatter()));
-        KSystemLogger.addLogHandler(new KTerminalLogHandler(new KSimpleLogFormatter()));
-
-        eventSystem.startPolling();
-        KStandardTestClass.context = new KProxiedEngineContext(
-            activator,
-            containerResolver,
-            objectRegistry,
-            eventSystem,
-            messageSystem,
-            resourceLoader,
-            assetLoader,
-            frameTaskSystem,
-            ctx -> {
-                eventSystem.stopPolling();
-            }
-        );
-        activator.addContext(KStandardTestClass.context);
-        KStandardTestClass.msgSystem = messageSystem;
-        KSystemLogger.activateFileLogging();
-        KSystemLogger.setLogLevel(KLogLevel.INFO);
-    }
 
     /**
      * Default constructor.
